+++ guillotina/db/storages/pg.py
@@ -44,7 +44,7 @@ try:
         "guillotina_db_pg_lock_time_seconds",
         "Histogram of time it takes to acquire locks (in seconds)",
         labelnames=["type"],
-        buckets=(0.0005, 0.001, 0.005, 0.01, 0.05, 0.1, 0.5, 1.0, 2.5, 5.0, INF,),
+        buckets=(0.0005, 0.001, 0.005, 0.01, 0.05, 0.1, 0.5, 1.0, 2.5, 5.0, INF),
     )
 
     class watch(metrics.watch):
@@ -69,7 +69,7 @@ try:
 
 
 except ImportError:
-    watch = metrics.watch  # type: ignore
+    watch = metrics.dummy_watch  # type: ignore
 
 log = glogging.getLogger("guillotina.storage")